feat: add login with jwt token

+++ src/main/java/com/example/demo/src/user/model/PostLoginRes.java
@@ -11,6 +11,6 @@ import lombok.Setter;
 @AllArgsConstructor
 public class PostLoginRes {
 
-    private int userIdx;
+    private Long userIdx;
     private String jwt;
 }
